// src/app/clubs_and_events_project/models/enums.ts
export enum CategoryClub {
  ACADEMY = 'ACADEMY',
  SPORTS = 'SPORTS',
  ARTS = 'ARTS',
  SOCIAL = 'SOCIAL',
  CULTURAL = 'CULTURAL'
}

export enum EventType {
  MEETING = 'MEETING',
  WORKSHOP = 'WORKSHOP',
  COMPETITION = 'COMPETITION'
}

export enum SpaceType {
  AUDITORIUM = 'Auditorium',
  GYMNASIUM = 'Gymnasium',
  STUDIO = 'Studio',
  LABORATORY = 'Laboratory',
  CLASSROOM = 'Classroom',
  CONFERENCE_ROOM = 'Conference Room',
  STUDY_ROOM = 'Study Room',
  SPORTS_FIELD = 'Sports Field',
  MEETING_ROOM = 'Meeting Room',
  DINING_HALL = 'Dining Hall',
  LIBRARY = 'Library',
  COMPUTER_LAB = 'Computer Lab',
  MUSIC_ROOM = 'Music Room',
  ART_GALLERY = 'Art Gallery',
  LECTURE_HALL = 'Lecture Hall'
}